Refactor the class in the badge namespace.
