+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Lyrics/Content/Components/Badges/LanguageBadge.cs
@@ -15,13 +15,13 @@ using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Components.UserInterfaceV2
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.States;
 using osu.Game.Rulesets.Karaoke.Utils;
 
-namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Content.Components.Badge;
+namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Content.Components.Badges;
 
-public partial class LanguageInfo : SubInfo, IHasPopover
+public partial class LanguageBadge : Badge, IHasPopover
 {
     private readonly Bindable<CultureInfo?> languageBindable;
 
-    public LanguageInfo(Lyric lyric)
+    public LanguageBadge(Lyric lyric)
         : base(lyric)
     {
         languageBindable = lyric.LanguageBindable.GetBoundCopy();
