+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Lyrics/Content/List/InfoControl.cs
@@ -18,7 +18,7 @@ using osu.Game.Rulesets.Karaoke.Edit.ChangeHandlers.Lyrics;
 using osu.Game.Rulesets.Karaoke.Edit.Utils;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Objects.Utils;
-using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Content.Components.Badge;
+using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Content.Components.Badges;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Content.Components.FixedInfo;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.States;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.States.Modes;
@@ -180,25 +180,25 @@ public partial class InfoControl : CompositeDrawable, IHasContextMenu
                     return null;
 
                 case LyricEditorMode.EditReferenceLyric:
-                    return new ReferenceLyricInfo(lyric);
+                    return new ReferenceLyricBadge(lyric);
 
                 case LyricEditorMode.EditLanguage:
-                    return new LanguageInfo(lyric);
+                    return new LanguageBadge(lyric);
 
                 case LyricEditorMode.EditRuby:
-                    return new LanguageInfo(lyric);
+                    return new LanguageBadge(lyric);
 
                 case LyricEditorMode.EditTimeTag:
                     return createTimeTagModeSubInfo(editorMode.GetEditStep<TimeTagEditStep>(), lyric);
 
                 case LyricEditorMode.EditRomanisation:
-                    return new LanguageInfo(lyric);
+                    return new LanguageBadge(lyric);
 
                 case LyricEditorMode.EditNote:
                     return null;
 
                 case LyricEditorMode.EditSinger:
-                    return new SingerInfo(lyric);
+                    return new SingerBadge(lyric);
 
                 default:
                     throw new ArgumentOutOfRangeException(nameof(editorMode));
@@ -209,11 +209,11 @@ public partial class InfoControl : CompositeDrawable, IHasContextMenu
                 switch (editMode)
                 {
                     case TimeTagEditStep.Create:
-                        return new LanguageInfo(lyric);
+                        return new LanguageBadge(lyric);
 
                     case TimeTagEditStep.Recording:
                     case TimeTagEditStep.Adjust:
-                        return new TimeTagInfo(lyric);
+                        return new TimeTagBadge(lyric);
 
                     default:
                         throw new ArgumentOutOfRangeException(nameof(editMode));
